requirements.txt
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
colorlog==6.9.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
persiantools==5.2.0
python-dotenv==1.1.0
python-telegram-bot==22.0
pytz==2025.2
requests==2.32.3
sniffio==1.3.1
tzdata==2025.2
urllib3==2.4.0
